+++ Erlang-Legacy/Assets/Core/Environment/EatFruitOfLifeTrigger.cs.meta
