+++ Erlang-Legacy/Assets/Core/Environment/OpenTreasureTrigger.cs.meta
